refactor: game page 에 사용하는 type 분리
