+++ src/components/boxscore/home-label.tsx
@@ -1,0 +1,7 @@
+export function HomeLabel() {
+  return (
+    <span className="rounded-sm bg-red-300 px-1 py-1 text-xs text-white">
+      홈
+    </span>
+  )
+}
