+++ src/types/index.ts
@@ -1,2 +1,3 @@
-export type { MenuData, LnbDataArray } from './header'
 export type { BannerData, TabsType, BannerInterface } from './banner'
+export type { GameSchedule, ScheduleType } from './game'
+export type { MenuData, LnbDataArray } from './header'
